backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
youtube-transcript-api==0.6.1
python-dotenv==1.0.1
